Move stackable check to parser.
Whether a lexeme is stackable is a concern for the parser/machine, not
the lexer!

While at it, make the 'lexeme' we pass to the parser/machine only the
dict of groups that matter. The parse shouldn't have to care about other
details, and it removes the lass Machine → Lexer call.